+++ src/entities/core-entities/market/market.ts
@@ -22,6 +22,13 @@ export class Market {
   @Expose({ name: 'MainLine' })
   mainLine?: string;
 
+  @Expose({ name: 'MarketStatus' })
+  marketStatus?: MarketStatus;
+
   @Expose({ name: 'Status' })
-  status?: MarketStatus;
+  set status(value: MarketStatus | undefined) {
+    if (value !== undefined && this.marketStatus === undefined) {
+      this.marketStatus = value;
+    }
+  }
 }